feat(editing): implement shift and split functions for lyrics
- Add shift functions to batch shift timing for words, lines, and paragraphs in lyrics.
- Implement split functions to divide words and lines based on character position or time.
- Create comprehensive tests for both shift and split functionalities to ensure correctness.
- Update index rebuilding logic to maintain data integrity after modifications.
- Refactor exports to organize editing API under features/editing.
