+++ src/features/editing/index.ts
@@ -5,6 +5,5 @@
 
 export { mergeLines, mergeWords } from "./merge";
 export { shiftLines, shiftParagraphs, shiftRange, shiftWords } from "./shift";
-export { splitLine, splitWord } from "./split";
 export type { SplitLineOptions, SplitWordOptions } from "./split";
-
+export { splitLine, splitWord } from "./split";
