+++ src/features/editing/split.ts
@@ -3,10 +3,10 @@
  * Split Word/Line into multiple parts
  */
 
-import { createWord } from "../factories/createWord";
-import type { ValidationResult } from "../schemas/result";
-import { failure, success } from "../schemas/result";
-import type { Line, Lyric, WordTimeline } from "../types";
+import { createWord } from "../../factories/createWord";
+import type { ValidationResult } from "../../schemas/result";
+import { failure, success } from "../../schemas/result";
+import type { Line, Lyric, WordTimeline } from "../../types";
 import { rebuildIndex, reindexPositions } from "./helpers";
 
 /**
