+++ src/index.ts
@@ -1,8 +1,8 @@
 // Core exports
 export * from "./Constants";
-// Editing API
-export * from "./editing";
 export * from "./factories";
+// Editing API
+export * from "./features/editing";
 // Schemas (Result types for error handling)
 export * from "./schemas/result";
 export * from "./types";
